Reorganize component css files in client src directory to clarify usage.
CSS files relevant to a specific React component have been placed in the corresponding subdirectory. For example, styling for elements only on the homepage is now in src/components/home/style.css

CSS files relevant to the entire website (e.g. container sizes, fonts, etc.) are in client/src/css, organized into SMACSS format.
