+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fd-client",
-  "version": "0.7.0",
+  "version": "0.7.1",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.3.0",
