+++ client/src/css/style.css
@@ -1,0 +1,4 @@
+.hero .ui.header {
+    font-size: 3em;
+    text-align: left;
+}
